set the project name to washdc and bump the cmake version up a bit
setting the project name is required so that `make install` works as
expected.

cmake version 3.6 is a lot newer than 2.8.4, but it's still very
conservative because 3.6 actually came out a few months before i made
the first commit to WashingtonDC.  so hopefully everybody's package
managers will have it by now.