implement Polygon culling.
this is currently done in software in the rendering backend due to the
PVR2's bias value, which does not have an OpenGL-counterpart.

in the future i hope to be able to implement this with a geometry
shader.  Another interesting possibility is to use OpenGL's own
culling capabilities when the bias value is 0.