infrastructure for logging PowerVR2 operations
for now, the proxy doesn't actually log anything so it's practically
worthless.  It also is only inserted in front of the TA and not the
rest of the PowerVR2.  Subsequent commits will address these
incomplete facets.

the point of this is to eventually save everything that happens to a
binary file and then have a program that can replay it, both on the
emulator and on a real dreamcast.  there will also be a program that
converts these dumps to a textual form.  this will greatly aid in
debugging and reverse-engineering eforts.