replace #ifdef branches in shaders with alternative shader linking
instead of implementing compile-time options to shaders by using the
glsl pre-processor, now we have multiple versions of each part of the
pipeline in different GLSL source files.

the reason for this is that i want to be able to run all of the
.vert/.frag files through glslangValidator.  that would have been
possible with the preprocessor, but we wouldn't have been able to test
every version.

Note that there are still further changes that need to be made before
these shader files are finally able to run through glslangValidator,
but this was the largest and most significant change.

Unfortunately this is going to be harder to read because you'll have
to refer to gfxgl[34]_renderer.c to figure out how the different .frag
and .vert files are linked together, but having shader code that
passes glslangValidator is more important than having shader code that is
easy to understand.